OneVision/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
requests>=2.31.0
pyttsx3>=2.90
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0